Create fr.lang (Arcanist for Necesse)
Hello ! Here’s a french translation for your mod. Feel free to integrate it if you have the time :)

I’m not sure of the right translation for "charloag". I tried to make it the equivalent of "Coal + log", so in French "Charbon + Bûche = Charbûche", is it right for you ?

For propick, I cheated as it is another Portmanteau word that is better in english. So it’s a "Professional pickaxe" in French and I made the "Propick" a rune in its description.